first pass of generated code

+++ openzt/src/capture_ztlog.rs
@@ -44,11 +44,12 @@ mod zoo_logging {
     use crate::{capture_ztlog::log_from_zt, util::get_string_from_memory};
     use openzt_detour::ZOOLOGGING_LOG;
 
+    // TODO: Generated signature uses i32/i8 types - verify if sign matters
     #[detour(ZOOLOGGING_LOG)]
-    unsafe extern "cdecl" fn zoo_log_func(source_file: u32, param_2: u32, param_3: u32, _param_4: u8, _param_5: u32, _param_6: u32, log_message: u32) {
+    unsafe extern "cdecl" fn zoo_log_func(source_file: u32, param_2: i32, param_3: i32, _param_4: i8, _param_5: i32, _param_6: i32, log_message: u32) {
         let source_file_string = get_string_from_memory(source_file);
         let log_message_string = get_string_from_memory(log_message);
-        log_from_zt(&source_file_string, param_2, param_3, &log_message_string);
+        log_from_zt(&source_file_string, param_2 as u32, param_3 as u32, &log_message_string);
     }
 }
